Changes from /ask
can you remind me of what is the role of "examples" folder here? I think it was supposed that the RustyFlow is a library, and we run example there. But are we using any example when we use train script and chat script? Let's rearrange a bit to make it clear.